pyproject.toml
[project]
name = "trip-planner-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "geopy>=2.4.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "supabase>=2.28.0",
    "tavily-python>=0.7.21",
    "uvicorn>=0.41.0",
]
